+++ CHANGELOG.md
@@ -7,8 +7,10 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 
 ## [Unreleased]
 
+## [0.3.0] - 2025-01-27
+
 ### Changed
-- Markdown renderer is on by defaul
+- Markdown renderer is on by default
 - When no sandbox, use host as fallback
 - Command approval now prints the command